requirements.txt
keybert
transformers
pyautogui
openai
streamlit

# LLM
python-dotenv
langchain
pinecone

# API Server
fastapi
uvicorn[standard]
python-multipart